# Conflicts:
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.ac/src/main/java/nts/uk/ctx/at/record/ac/workschedule/WorkScheduleWorkInforAcFinder.java
#	nts.uk/uk.at/at.ctx/schedule/nts.uk.ctx.at.schedule.pub/src/main/java/nts/uk/ctx/at/schedule/pub/schedule/workschedule/WorkScheduleExport.java

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/outsideot/overtime/KshstOverTimePK.java
@@ -44,7 +44,7 @@ public class KshstOverTimePK implements Serializable {
      * @param cid the cid
      * @param overTimeNo the over time no
      */
-    public KshstOverTimePK(String cid, short overTimeNo) {
+    public KshstOverTimePK(String cid, int overTimeNo) {
         this.cid = cid;
         this.overTimeNo = overTimeNo;
     }
